backend/auth/requirements.txt
psycopg2-binary==2.9.9
google-auth==2.25.2